feat: update command
